+++ CHANGES.md
@@ -1,3 +1,6 @@
+- Flush rewrite rules on activation and deactivation of plugin
+- Added contactadvertentie-category to permalink structure 
+
 0.2
 - Improved capabilities and role commits
 - Renamed Class to EJO_Contactads